Add files via upload
In the project's code, the AES encryption and decryption are implemented using the PyCryptodome library, which is a Python package providing cryptographic functionalities. Specifically, the AES encryption mode used in the project is Galois/Counter Mode (GCM). GCM is an authenticated encryption mode that provides both confidentiality (encryption) and integrity (authentication) of the encrypted data.

In the project, the AES encryption and decryption operations are performed using the AES.new() constructor provided by the PyCryptodome library, with the AES.MODE_GCM parameter specifying the GCM mode of operation. This ensures that the sensitive data, such as passwords, is securely encrypted and decrypted using a strong cryptographic algorithm and mode of operation.